Added transaction routes, controller, services, and repositories

+++ Backend/repositories/transaction_repository.js
@@ -12,6 +12,10 @@ export const findPurchasesByBuyer = async (buyer) => {
     return await Transaction.find({ buyer });
 };
 
+export const findPurchasesByPhotographer = async (photograper) => {
+    return await Transaction.find({ photographer });
+};
+
 export const findPurchasesByPhoto = async (photo) => {
     return await Transaction.find({ photo });
 };
@@ -25,6 +29,10 @@ export const findPurchase = async (buyer, photo) => {
 };
 
 export const deletePurchase = async (id) => {
+     const purchase=await findById(id)
+    if (purchase.status !== "pending") {
+        throw new Error("Only pending purchases can be cancelled.");
+    }
     return await Transaction.findByIdAndDelete(id);
 };
